Simplify sending multi-cursor selections
Eliminated the complex mechanism needed to find multi-cursor
selections and insert proper tracebacks.  In place of this
mechanism a string representing the full file length is sent,
but areas which are not selected will only contain blank lines.
As python ignores extra blank lines, this should be transparent
to the user.  This is both simpler and more robust.